requirements.txt
django-environ==0.10.0
pymongo==4.15.5
redis==7.1.0
loguru==0.7.3